OSIS-724 Correction pour PR et refactoring du Business
