+++ base/tests/business/test_learning_unit_creation.py
@@ -24,6 +24,8 @@
 #
 ##############################################################################
 from django.test import TestCase
+
+import base.business.learning_units.simple.creation
 from base.tests.factories.academic_year import create_current_academic_year
 from base.business import learning_unit
 from base.tests.factories.learning_container_year import LearningContainerYearFactory
@@ -52,7 +54,7 @@ class LearningUnitCreationTest(TestCase):
     def test_create_learning_unit_year(self):
         data_dict = self.get_data_dict(learning_container_year_types.OTHER_COLLECTIVE)
         data = data_dict.get('data')
-        new_learning_unit_yr = learning_unit.create_learning_unit_year(data_dict)
+        new_learning_unit_yr = base.business.learning_units.simple.creation.create_learning_unit_year(data_dict)
         self.assertEqual(new_learning_unit_yr.academic_year, data_dict.get('academic_year'))
 
         self.assertEqual(new_learning_unit_yr.learning_unit, data_dict.get('new_learning_unit'))
@@ -72,17 +74,17 @@ class LearningUnitCreationTest(TestCase):
         learning_container_yr = LearningContainerYearFactory(academic_year=self.current_academic_year,
                                                              learning_container=learning_container)
 
-        new_learning_component_yr = learning_unit\
-            .create_learning_component_year(learning_container_yr,
-                                            learning_unit.DEFAULT_ACRONYM_LECTURING_COMPONENT,
-                                            learning_component_year_type.LECTURING)
+        new_learning_component_yr = base.business.learning_units.simple.creation.create_learning_component_year(learning_container_yr,
+                                                                                                                base.business.learning_units.simple.creation.DEFAULT_ACRONYM_LECTURING_COMPONENT,
+                                                                                                                learning_component_year_type.LECTURING)
         self.assertEqual(new_learning_component_yr.learning_container_year, learning_container_yr)
-        self.assertEqual(new_learning_component_yr.acronym, learning_unit.DEFAULT_ACRONYM_LECTURING_COMPONENT)
-        self.assertEqual(new_learning_component_yr.type, learning_unit.learning_component_year_type.LECTURING)
+        self.assertEqual(new_learning_component_yr.acronym,
+                         base.business.learning_units.simple.creation.DEFAULT_ACRONYM_LECTURING_COMPONENT)
+        self.assertEqual(new_learning_component_yr.type, learning_component_year_type.LECTURING)
 
     def test_create_with_lecturing_and_practical_components(self):
         data_dict = self.get_data_dict(learning_container_year_types.OTHER_COLLECTIVE)
-        learning_unit.create_with_lecturing_and_practical_components(data_dict)
+        base.business.learning_units.simple.creation.create_with_lecturing_and_practical_components(data_dict)
 
         self.assertEqual(mdl_base.learning_component_year.LearningComponentYear.objects.all().count(), 2)
         self.assertEqual(mdl_base.learning_component_year.LearningComponentYear.objects
@@ -102,17 +104,17 @@ class LearningUnitCreationTest(TestCase):
 
     def test_create_with_untyped_component(self):
         data_dict = self.get_data_dict(learning_container_year_types.OTHER_COLLECTIVE)
-        learning_unit.create_with_untyped_component(data_dict)
+        base.business.learning_units.simple.creation.create_with_untyped_component(data_dict)
         self.assertEqual(mdl_base.learning_component_year.LearningComponentYear.objects.all().count(), 1)
         self.assertEqual(mdl_base.learning_component_year.LearningComponentYear.objects
-                         .filter(acronym=learning_unit.UNTYPED_ACRONYM).count(), 1)
+                         .filter(acronym=base.business.learning_units.simple.creation.UNTYPED_ACRONYM).count(), 1)
         self.assertEqual(mdl_base.learning_component_year.LearningComponentYear.objects
                          .filter(type__isnull=True).count(), 1)
         self.assertEqual(mdl_base.entity_component_year.EntityComponentYear.objects.all().count(), 1)
         self.assertEqual(mdl_base.learning_unit_year.LearningUnitYear.objects.all().count(), 1)
         self.assertEqual(mdl_base.learning_unit_component.LearningUnitComponent.objects.all().count(), 1)
 
-    @mock.patch("base.business.learning_unit.create_with_lecturing_and_practical_components")
+    @mock.patch("base.business.learning_units.simple.creation.create_with_lecturing_and_practical_components")
     def test_create_learning_unit_content_create_with_lecturing_and_practical_components(self, mock):
         container_type_with_default_component = [learning_container_year_types.COURSE,
                                                  learning_container_year_types.MASTER_THESIS,
@@ -120,17 +122,17 @@ class LearningUnitCreationTest(TestCase):
                                                  learning_container_year_types.INTERNSHIP]
         for container_type in container_type_with_default_component:
             data_dict = self.get_data_dict(container_type)
-            learning_unit.create_learning_unit_content(data_dict)
+            base.business.learning_units.simple.creation.create_learning_unit_content(data_dict)
             self.assertTrue(mock.called)
 
-    @mock.patch("base.business.learning_unit.create_with_untyped_component")
+    @mock.patch("base.business.learning_units.simple.creation.create_with_untyped_component")
     def test_create_learning_unit_content_create_with_untyped_component(self, mock):
         container_type_without_default_component = [learning_container_year_types.DISSERTATION,
                                                     learning_container_year_types.OTHER_INDIVIDUAL,
                                                     learning_container_year_types.EXTERNAL]
         for container_type in container_type_without_default_component:
             data_dict = self.get_data_dict(container_type)
-            learning_unit.create_learning_unit_content(data_dict)
+            base.business.learning_units.simple.creation.create_learning_unit_content(data_dict)
             self.assertTrue(mock.called)
 
     def get_data_dict(self, container_type):
