--- base/views/learning_units/proposal/__init__.py
@@ -1,1 +1,0 @@
-from base.views.learning_units.proposal import creation
